refactor(blorktools): material creation util
