--- packages/blorktools/src/asset_debugger/index.js
@@ -1,10 +1,0 @@
-/**
- * Asset Debugger - Main Entry Point
- * Exports all necessary functions and components for the asset debugger
- */
-
-// Export utility functions
-export { formatFileSize } from './util/materials-util.js';
-
-// Export state management functions
-export { getState, updateState, initDraftState } from './scene/state.js';
