+++ packages/blorktools/src/asset_debugger/modals/settings-modal/settings-modal.js
@@ -1,7 +1,7 @@
 /**
  * Settings Modal UI Component for Asset Debugger
  */
-import { saveSettings, loadSettings, getDefaultSettings } from '../../util/localstorage-util.js';
+import { saveSettings, loadSettings, getDefaultSettings } from '../../util/data/localstorage-util.js';
 import { updateRigOptions } from '../../util/rig/rig-manager.js';
 
 export class SettingsModal {
